projects/meeting-creation-automation/meeting-creation-automation-main/requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
email-validator==2.1.0
python-dateutil==2.8.2
icalendar==5.0.11
flask==2.3.3
